+++ src/activities/components/ActivityList.vue
@@ -48,6 +48,144 @@
         {{ filteredActivities.length }} / {{ activities.length }}
       </div>
     </div>
+    <div
+      v-if="displayedActivitiesGroupedByDate.length"
+      class="row no-wrap items-center justify-end q-px-sm q-mt-sm"
+    >
+      <QBtn
+        color="secondary"
+        icon="event"
+        padding="4px 13px"
+        rounded
+        size="sm"
+        unelevated
+        :label="$t('ACTIVITYLIST.ICS_DIALOG.TITLE')"
+        @click="icsDialog = true"
+      />
+      <CustomDialog
+        v-model="icsDialog"
+        width="500px"
+        actions-align="between"
+      >
+        <template #title>
+          {{ $t('ACTIVITYLIST.ICS_DIALOG.TITLE') }}
+        </template>
+        <template #message>
+          <p>
+            {{ $t('ACTIVITYLIST.ICS_DIALOG.INTRO') }}
+          </p>
+
+          <QList>
+            <QItem class="q-pl-none">
+              <QItemSection
+                avatar
+                top
+              >
+                <QAvatar
+                  color="primary"
+                  text-color="white"
+                  icon="file_download"
+                />
+              </QItemSection>
+
+              <QItemSection>
+                <QItemLabel>{{ $t('ACTIVITYLIST.ICS_DIALOG.DOWNLOAD_TITLE') }}</QItemLabel>
+                <QItemLabel caption>
+                  {{ $t('ACTIVITYLIST.ICS_DIALOG.DOWNLOAD_TEXT') }}
+                </QItemLabel>
+                <QItemLabel>
+                  <QBtn
+                    color="primary"
+                    class="q-mt-xs"
+                    type="a"
+                    :href="icsUrl"
+                    :label="$t('ACTIVITYLIST.ICS_DIALOG.DOWNLOAD_BUTTON_LABEL')"
+                  />
+                </QItemLabel>
+              </QItemSection>
+            </QItem>
+
+            <QSeparator
+              spaced
+              style="margin-left: 56px;"
+            />
+
+            <QItem class="q-pl-none">
+              <QItemSection
+                avatar
+                top
+              >
+                <QAvatar
+                  color="primary"
+                  text-color="white"
+                  icon="sync"
+                />
+              </QItemSection>
+
+              <QItemSection>
+                <QItemLabel>{{ $t('ACTIVITYLIST.ICS_DIALOG.SUBSCRIBE_TITLE') }}</QItemLabel>
+                <QItemLabel caption>
+                  {{ $t('ACTIVITYLIST.ICS_DIALOG.SUBSCRIBE_TEXT') }}
+                </QItemLabel>
+                <QItemLabel>
+                  <QField
+                    filled
+                    class="q-mt-xs"
+                  >
+                    <template #append>
+                      <QBtn
+                        flat
+                        rounded
+                        icon="fas fa-copy"
+                        @click="copyLink"
+                      >
+                        <q-tooltip
+                          anchor="top middle"
+                          self="bottom middle"
+                        >
+                          {{ $t('URL_CLICK_TO_COPY') }}
+                        </q-tooltip>
+                      </QBtn>
+                    </template>
+                    <template #control>
+                      <!-- https://css-tricks.com/snippets/css/prevent-long-urls-from-breaking-out-of-container/ -->
+                      <div
+                        class="self-center full-width no-outline"
+                        style="word-break: break-word; overflow-wrap: break-word;"
+                      >
+                        {{ icsUrl }}
+                      </div>
+                    </template>
+                  </QField>
+                </QItemLabel>
+              </QItemSection>
+            </QItem>
+          </QList>
+        </template>
+        <template #actions>
+          <a
+            rel="noopener noreferrer"
+            href="https://fileinfo.com/extension/ics"
+            target="_blank"
+            class="text-secondary q-pl-sm"
+          >
+            <QIcon
+              name="link"
+              size="xs"
+              class="q-mr-xs"
+            />
+            {{ $t('ACTIVITYLIST.ICS_DIALOG.MORE_INFOS') }}
+          </a>
+          <QBtn
+            v-close-popup
+            flat
+            color="primary"
+            autofocus
+            :label="$t('BUTTON.CLOSE')"
+          />
+        </template>
+      </CustomDialog>
+    </div>
     <KSpinner v-show="pending" />
     <div
       v-if="!pending && noActivitiesDueToFilters"
@@ -115,33 +253,44 @@
 <script>
 import ActivityItem from './ActivityItem'
 import KSpinner from '@/utils/components/KSpinner'
+import CustomDialog from '@/utils/components/CustomDialog'
 import bindRoute from '@/utils/mixins/bindRoute'
 
 import {
+  copyToClipboard,
+  QAvatar,
+  QField,
   QSelect,
   QInfiniteScroll,
   QItem,
   QItemSection,
   QItemLabel,
   QIcon,
+  QList,
   QBanner,
   QBtn,
+  QSeparator,
 } from 'quasar'
 
 const NUM_ACTIVITIES_PER_LOAD = 25
 
 export default {
   components: {
+    CustomDialog,
+    QAvatar,
     QInfiniteScroll,
     ActivityItem,
     KSpinner,
+    QField,
     QSelect,
     QItem,
     QItemSection,
     QItemLabel,
     QIcon,
+    QList,
     QBanner,
     QBtn,
+    QSeparator,
   },
   mixins: [
     bindRoute({
@@ -174,9 +323,14 @@ export default {
       type: Array,
       default: () => [],
     },
+    icsUrl: {
+      type: String,
+      default: null,
+    },
   },
   data () {
     return {
+      icsDialog: false,
       numDisplayed: NUM_ACTIVITIES_PER_LOAD,
       types: [],
     }
@@ -270,6 +424,13 @@ export default {
     clearFilters () {
       this.slots = 'all'
       this.type = 'all'
+    },
+    copyLink () {
+      return copyToClipboard(this.icsUrl).then(() => {
+        this.$store.dispatch('toasts/show', {
+          message: 'URL_COPIED_TOAST',
+        }, { root: true })
+      })
     },
   },
 }