+++ src/utils/components/CustomDialog.vue
@@ -3,14 +3,17 @@
     v-bind="$attrs"
     v-on="$listeners"
   >
-    <QCard class="q-dialog-plugin">
+    <QCard
+      class="q-dialog-plugin"
+      :style="{ width: $attrs.width ? $attrs.width : ''}"
+    >
       <QCardSection class="q-dialog__title">
         <slot name="title" />
       </QCardSection>
       <QCardSection class="q-dialog__message">
         <slot name="message" />
       </QCardSection>
-      <QCardActions align="right">
+      <QCardActions :align="$attrs['actions-align'] ? $attrs['actions-align'] : 'right'">
         <slot name="actions" />
       </QCardActions>
     </QCard>